refactor : PanoramaFavorite 기능 리펙토링 및 컨벤션 수정 (#21)
