+++ src/favorite/dto/panoramaFavorite/panoramaFavorite-request.dto.ts
@@ -1,0 +1,11 @@
+import { IsNotEmpty } from "class-validator";
+
+export class PanoramaFavoriteRequestDto {
+
+    @IsNotEmpty()
+    userID: number;
+
+    @IsNotEmpty()
+    panoramaViewerID: number;
+    
+}
