--- src/panorama/dto/panoramaFavorite/panoramaFavorite-request.dto.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from "class-validator";
-
-export class PanoramaFavoriteRequestDto {
-    @IsNotEmpty()
-    userID: number;
-
-    @IsNotEmpty()
-    panoramaViewerID: number;
-}
