+++ src/panorama/entities/panoramaImage.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
-@Entity('panoramaimage')
+@Entity()
 export class PanoramaImage {
     @PrimaryGeneratedColumn()
     panoramaViewerImageID: number;
