+++ src/panorama/panorama.module.ts
@@ -3,7 +3,6 @@ import { PanoramaViewerController } from './controller/panoramaViewer.controller
 import { PanoramaViewerService } from './service/panoramaViewer.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PanoramaViewer } from './entities/panoramaViewer.entity';
-import { PanoramaFavorite } from './entities/panoramaViewerFavorite.entity';
 import { PanoramaImage } from './entities/panoramaImage.entity';
 
 @Module({
@@ -11,7 +10,6 @@ import { PanoramaImage } from './entities/panoramaImage.entity';
     TypeOrmModule.forFeature([
       PanoramaImage,
       PanoramaViewer,
-      PanoramaFavorite
     ])
   ],
   controllers: [
